+++ examples/Serialization/create_from_config.py
@@ -78,7 +78,7 @@ CONFIG = str(
         "recommender": {
             "type": "SequentialGreedyRecommender",
             "surrogate_model": {
-                "type": "GaussianProcessModel"
+                "type": "GaussianProcessSurrogate"
             },
             "acquisition_function_cls": "qEI"
         },